backend/requirements.txt
fastapi~=0.115.8
pydantic~=2.10.6
pyserial~=3.5
python-dotenv~=1.0.1
